* Minor improvements. #45
